IaC Webhook Service OutboundMessage methods ConfigStackSettings so we can use embedded applicationjson files!!!

+++ src/RevolutionaryStuff.Core/Services/TemporaryStreamFactory/ITemporaryStreamFactory.cs
@@ -4,5 +4,5 @@ namespace RevolutionaryStuff.Core.Services.TemporaryStreamFactory;
 
 public interface ITemporaryStreamFactory
 {
-    Stream Create(int? capacity = null);
+    Stream Create(long? capacity = null);
 }
